+++ src/core/db/pdo.php
@@ -1,11 +1,16 @@
 <?php
     namespace Leaf\Core\Db;
 
+    /**
+	 * Leaf Core PDO
+	 * -----------------------
+	 * Leaf's adaptation of **PDO**
+	 */
     class PDO {
         protected $connection;
         protected $queryResult;
 
-        public function connectPDO($host, $dbname, $user, $password) {
+        public function connect($host, $dbname, $user, $password) {
             try {
                 $connection = new PDO("mysql:host=$host;dbname=$dbname", $user, $password);
                 $connection->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
@@ -18,27 +23,55 @@
 
         public function query(string $query, array $params = []) {
             if ($this->connection == null) {
-                echo "Initialise your database first with connectMysqli()";
+                echo "Initialise your database first with connect()";
+                exit();
             }
             
+            if(!$params) {
+                $this->queryResult = $this->connection->query($query);
+            } else {
+                $stmt = $this->connection->prepare($query);
+                $stmt->bindParam(...$params);
+                $this->queryResult = $stmt->execute();
+            }
             
             return $this;
 		}
 		
-		public function select() {
-			// 
+		public function select(string $table, string $items = "*", string $options = "", array $params = []) {
+			if (strlen($options) > 1) {
+				$this->query("SELECT $items FROM $table WHERE $options", $params);
+			} else {
+				$this->query("SELECT $items FROM $table", $params);
+			}
+			
+			return $this;
 		}
 
-		public function update() {
-			// 
+		public function delete(string $table, string $options = "", array $params = []) {
+			if (strlen($options) > 1) {
+				$this->query("DELETE FROM $table WHERE $options", $params);
+			} else {
+				$this->query("DELETE FROM $table", $params);
+			}
+			
+			return $this;
 		}
 
-		public function delete() {
-			// 
+		public function insert(string $table, string $column, string $value, array $params = []) {
+			$this->query("INSERT INTO $table ($column) VALUES ($value)", $params);
+			
+			return $this;
 		}
 
-		public function insert() {
-			// 
+		public function update(string $table, string $updateOptions, string $options, array $params = []) {
+			if (strlen($options) > 1) {
+				$this->query("UPDATE $table SET $updateOptions WHERE $options", $params);
+			} else {
+				$this->query("UPDATE $table SET $updateOptions", $params);
+			}
+			
+			return $this;
 		}
 
 		public function count() {
@@ -46,15 +79,21 @@
 		}
 
 		public function fetchObj() {
-			// 
+			return $this->queryResult->fetch(PDO::FETCH_OBJ);
 		}
 
         public function fetchAssoc() {
-            // 
+            return $this->queryResult->fetch(PDO::FETCH_ASSOC);
         }
 
-        public function fetchAll() {
-            // 
+        public function fetchAll($type = FETCH_OBJ) {
+            if ($type == "obj" || $type == "object" || $type == FETCH_OBJ) {
+				$type = FETCH_OBJ;
+			}
+			if ($type != "obj" && $type != FETCH_OBJ || $type == "assoc") {
+				$type = FETCH_ASSOC;
+            }
+            $this->queryResult->fetchAll(PDO::$type);
         }
 
         public function result() {